fix client async issue
